dom/tests/mochitest/pointerlock/file_removedFromDOM.html

<!DOCTYPE HTML>
<html>
  <!--
  https://bugzilla.mozilla.org/show_bug.cgi?id=633602

  Test DOM tree in full screen
  -->
  <head>
    <title>Bug 633602 - file_DOMtree.html</title>
    <script type="application/javascript" src="/tests/SimpleTest/EventUtils.js">
    </script>
    <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js">
    </script>
    <script type="application/javascript" src="pointerlock_utils.js"></script>
    <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
    <style>
    </style>
  </head>
  <body>
    <a target="_blank"
       href="https://bugzilla.mozilla.org/show_bug.cgi?id=633602">
      Mozilla Bug 633602
    </a>
    <div id="div"></div>
    <pre id="test">
      <script type="text/javascript">
        /*
         * Test for Bug 633602
         * Checks if pointer is unlocked when element is removed from
         * the DOM Tree
         */

        SimpleTest.waitForExplicitFinish();

        var div = document.getElementById("div")
          , removedDOM = false;

        function runTests() {
          ok(removedDOM, "Pointer should be unlocked when " +
            "an element is removed the DOM Tree");
        }

        document.addEventListener("mozpointerlockchange", function (e) {
          if (document.mozPointerLockElement === div) {
            document.body.removeChild(div);
            removedDOM = !document.mozPointerLockElement;
            document.mozCancelFullScreen();
          }

        }, false);

        document.addEventListener("mozfullscreenchange", function (e) {
          if (document.mozFullScreenElement === div) {
              div.mozRequestPointerLock();
          }
          else {
            runTests();
            SimpleTest.finish();
          }
        }, false);

        function start() {
          div.mozRequestFullScreen();
        }
      </script>
    </pre>
  </body>
</html>
